Allow S3OutputStream to commit as part of close()
The existing RecordWriterProvider classes in this project need to be
careful to flush their writers before calling commit() on the
S3OutputStream and include comments warning about this need.

This PR moves the commit logic directly into the close() method
so that closing a wrapper can more elegantly initiate the multipart
upload.

I'm particularly interested in this change as it paves the way
for an elegant implementation of compression in this project.
In particular, the current implementation makes it necessary to
call GZIPOutputStream.finish() before closing, which is a method
specific to DeflatorOutputStream. If commit can happen inside close(),
then we can handle compressors more generically via the OutputStream
interface.