Dynamically create TopicPartitionWriters
Currently, this connector generates errors when configured with a
Transformer such RegexRouter or TimestampRouter which changes the
topic name.

In order to support topic Router transforms, we need to dynamically
create TopicPartitionWriters as we see new topic names, since the
existing logic creates a static set of writers for only the specific
topic partitions in the connector's assignment.