CC-3035: Upgrade the S3Mock test library
Upgrade the S3Mock library from 0.1.5 (on Scala 2.11) to 0.2.5 (on Scala 2.12), which incorporates a number of bug fixes. This is to take advantage of a new `S3Mock.shutdown()` method that properly closes the Akka framework and HTTP listener created in the S3Mock object.

For some reason, these still-bound HTTP listeners are not a problem on Java 8, but with Java 11 they prevent the next test from binding to the address and cause the test to fail.

With this commit, the build passes on both Java 8 and Java 11.