CC-6766 Handle null-valued records in all three output types (#280)
* Handle null-valued records in all three output types

The new code will give consistent handling of null-valued records.

Before the change, json would correctly serialize null, while
bytes and avro would hard-fail. The recommended solution was to
use SMT to filter null-valued records, but this fix will integrate
the "skip" functionality into the connector. It does not preclude
the use of the SMT to get more general features such as DLQ.