Close only requested partitions
Closing other partitions may look safe, but some partitions can in
theory remain closed even if they should be open.

For example topics-partitions A-0, A-1, B-0, B-1 may be assigned to two
tasks (first would have A-0,B-0; second A-1,B-1). When a new task comes,
it may be decided it gets B-0.
The first task calls close([B-0]); open([]) so that B-0 is handled by
the new task. First task should still be able to handle A-0.

Besides this solution, we could also open all assigned topics-partitions
on every open() call. That would work, but it's not the most effective
way.