MINOR: Replace Kafka Utils.join() with JDK API
The Kafka commons Utils class removed the "join" method in https://github.com/apache/kafka/pull/15823 in Kafka 3.8. Due to this reason the connector stopped working on Kafka Connect 3.8 as well.

Following the example in the upstream PR, I'm replacing the usage of Utils.join with the JDK API of String.join, and updating the code accordingly.

This fixes https://github.com/confluentinc/kafka-connect-storage-cloud/issues/781