Enable conditional writes to connector
Conditional write is enabled for only connectors running with at least once semantics using scheduled rotation.
The behaviour can be controlled using a configuration property.
When an upload fails due to file being present in s3, the connector scans for the next available file name in s3 which the connector can commit to and resets the offset to same