MINOR: Replace Kafka Utils.join() with JDK API (#810)
The Kafka commons Utils class removed the "join" method in
apache/kafka#15823 in Kafka 3.8. Due to this reason the connector
stopped working on Kafka Connect 3.8 as well.

Following the example in the upstream PR, I'm replacing the usage of
Utils.join with the JDK API of String.join, and updating the code
accordingly.

This is a new version of PR 804, but rebased for 10.0.x. Fixes #781